requirements.txt
pillow>=10.0.0
pycryptodome>=3.20.0